+++ ofrak_core/ofrak/core/entropy/entropy.py
@@ -17,6 +17,7 @@ LOGGER = logging.getLogger(__name__)
 
 try:
     from ofrak.core.entropy.entropy_c import get_entropy_c
+
     entropy_func = get_entropy_c()
 except:
     from ofrak.core.entropy.entropy_py import entropy_py as entropy_func